Maps implementation transferred to SelectLocationFragment - Location tracking implemented
