+++ starter/app/src/debug/res/values/google_maps_api.xml
@@ -3,5 +3,5 @@
         TODO: Before you run your application, you need a Google Maps API key.
     -->
     <string name="google_maps_key" templateMergeStrategy="preserve"
-        translatable="false"></string>
+        translatable="false">AIzaSyC6hBgGRJHLlKoosvbR-nPUg0w5Lg3cbWs</string>
 </resources>
