+++ starter/app/src/main/java/com/udacity/project4/locationreminders/reminderslist/ReminderListFragment.kt
@@ -1,6 +1,5 @@
 package com.udacity.project4.locationreminders.reminderslist
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.*
 import androidx.databinding.DataBindingUtil
@@ -9,7 +8,6 @@ import com.udacity.project4.R
 import com.udacity.project4.base.BaseFragment
 import com.udacity.project4.base.NavigationCommand
 import com.udacity.project4.databinding.FragmentRemindersBinding
-import com.udacity.project4.maps.MapsActivity
 import com.udacity.project4.utils.setDisplayHomeAsUpEnabled
 import com.udacity.project4.utils.setTitle
 import com.udacity.project4.utils.setup
@@ -71,9 +69,6 @@ class ReminderListFragment : BaseFragment() {
                 // Logout implementation
                 AuthUI.getInstance().signOut(requireContext())
                 requireActivity().onBackPressedDispatcher.onBackPressed()
-            }
-            R.id.show_maps -> {
-                startActivity(Intent(requireActivity(), MapsActivity::class.java))
             }
         }
         return super.onOptionsItemSelected(item)
