+++ starter/app/src/main/java/com/udacity/project4/locationreminders/savereminder/selectreminderlocation/SelectLocationFragment.kt
@@ -1,20 +1,55 @@
 package com.udacity.project4.locationreminders.savereminder.selectreminderlocation
 
+import android.Manifest
+import android.annotation.SuppressLint
+import android.content.pm.PackageManager
+import android.content.res.Resources
 import android.os.Bundle
+import android.util.Log
 import android.view.*
+import androidx.core.app.ActivityCompat
 import androidx.databinding.DataBindingUtil
+import com.google.android.gms.location.FusedLocationProviderClient
+import com.google.android.gms.location.Granularity
+import com.google.android.gms.location.LocationRequest
+import com.google.android.gms.location.LocationServices
+import com.google.android.gms.location.LocationServices.getFusedLocationProviderClient
+import com.google.android.gms.location.LocationSettingsRequest
+import com.google.android.gms.location.Priority
+import com.google.android.gms.maps.CameraUpdateFactory
+import com.google.android.gms.maps.GoogleMap
+import com.google.android.gms.maps.OnMapReadyCallback
+import com.google.android.gms.maps.SupportMapFragment
+import com.google.android.gms.maps.model.*
 import com.udacity.project4.R
 import com.udacity.project4.base.BaseFragment
 import com.udacity.project4.databinding.FragmentSelectLocationBinding
 import com.udacity.project4.locationreminders.savereminder.SaveReminderViewModel
 import com.udacity.project4.utils.setDisplayHomeAsUpEnabled
 import org.koin.android.ext.android.inject
+import java.util.*
 
-class SelectLocationFragment : BaseFragment() {
+class SelectLocationFragment : BaseFragment(), OnMapReadyCallback {
+
+    companion object {
+        private val TAG = SelectLocationFragment::class.java.simpleName
+        private const val REQUEST_LOCATION_PERMISSION = 1
+        private const val UPDATE_INTERVAL = (10 * 1000).toLong() // 10 secs
+        private const val FASTEST_INTERVAL: Long = 2000 // 2 secs
+    }
 
     // Use Koin to get the view model of the SaveReminder
     override val _viewModel: SaveReminderViewModel by inject()
+
+    private lateinit var map: GoogleMap
+    private lateinit var locationClient: FusedLocationProviderClient
+    private lateinit var locationRequest: LocationRequest
+
     private lateinit var binding: FragmentSelectLocationBinding
+
+    //--------------------------------------------------
+    // Lifecycle Methods
+    //--------------------------------------------------
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
@@ -28,14 +63,123 @@ class SelectLocationFragment : BaseFragment() {
         setHasOptionsMenu(true)
         setDisplayHomeAsUpEnabled(true)
 
-        // TODO: add the map setup implementation
-        // TODO: zoom to the user location after taking his permission
-        // TODO: add style to the map
-        // TODO: put a marker to location that the user selected
+        // Add the map setup implementation.
+        // Obtain the SupportMapFragment and get notified when the map is ready to be used.
+        initMap()
 
-        // TODO: call this function after the user confirms on the selected location
-        onLocationSelected()
+        // TODO: Call this function after the user confirms on the selected location
+        binding.fab.setOnClickListener {
+//            onLocationSelected()
+            connectToGetLocation()
+        }
+
         return binding.root
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        locationClient.flushLocations()
+//        locationClient.removeLocationUpdates(this)
+    }
+
+    //--------------------------------------------------
+    // Menu Methods
+    //--------------------------------------------------
+
+    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
+        inflater.inflate(R.menu.map_options, menu)
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem) = when (item.itemId) {
+        // Change the map type based on the user's selection.
+        R.id.normal_map -> {
+            map.mapType = GoogleMap.MAP_TYPE_NORMAL
+            true
+        }
+        R.id.hybrid_map -> {
+            map.mapType = GoogleMap.MAP_TYPE_HYBRID
+            true
+        }
+        R.id.satellite_map -> {
+            map.mapType = GoogleMap.MAP_TYPE_SATELLITE
+            true
+        }
+        R.id.terrain_map -> {
+            map.mapType = GoogleMap.MAP_TYPE_TERRAIN
+            true
+        }
+        else -> super.onOptionsItemSelected(item)
+    }
+
+    //--------------------------------------------------
+    // Location Methods
+    //--------------------------------------------------
+
+    @SuppressLint("MissingPermission")
+    private fun startLocationUpdates() {
+        // Create the location request to start receiving updates
+        // https://tomas-repcik.medium.com/locationrequest-create-got-deprecated-how-to-fix-it-e4f814138764
+        locationRequest = LocationRequest.Builder(
+            Priority.PRIORITY_HIGH_ACCURACY, UPDATE_INTERVAL
+        ).apply {
+            setMinUpdateDistanceMeters(5F)
+            setGranularity(Granularity.GRANULARITY_PERMISSION_LEVEL)
+            setWaitForAccurateLocation(true)
+            setMinUpdateIntervalMillis(FASTEST_INTERVAL)
+        }.build()
+
+        // Create LocationSettingsRequest object using location request
+        val builder = LocationSettingsRequest.Builder()
+        builder.addLocationRequest(locationRequest)
+        val locationSettingsRequest = builder.build()
+
+        // Check whether location settings are satisfied
+        // https://developers.google.com/android/reference/com/google/android/gms/location/SettingsClient
+        val settingsClient = LocationServices.getSettingsClient(requireActivity())
+        settingsClient.checkLocationSettings(locationSettingsRequest)
+    }
+
+    /*
+    private fun onLocationChanged(location: Location) {
+        // New location has now been determined
+        val msg = "Updated Location: " + location.latitude.toString() + "," + location.longitude.toString()
+        Toast.makeText(requireContext(), msg, Toast.LENGTH_SHORT).show()
+        // You can now create a LatLng Object for use with maps
+        val latLng = LatLng(location.latitude, location.longitude)
+    }
+     */
+
+    @SuppressLint("MissingPermission")
+    private fun connectToGetLocation() {
+        /*
+        locationClient = getFusedLocationProviderClient(requireActivity())
+        locationClient.requestLocationUpdates(
+            locationRequest, object : LocationCallback() {
+                override fun onLocationResult(locationResult: LocationResult) {
+                    // do work here
+                    locationResult.lastLocation?.let { onLocationChanged(it) }
+                }
+            },
+            Looper.myLooper()
+        )
+         */
+
+        locationClient = getFusedLocationProviderClient(requireActivity())
+        locationClient.lastLocation
+            .addOnSuccessListener { location -> // GPS location can be null if GPS is switched off
+                location?.let { goToLocation(location.latitude, location.longitude) }
+            }
+            .addOnFailureListener { e ->
+                Log.d("MapDemoActivity", "Error trying to get last GPS location")
+                e.printStackTrace()
+            }
+    }
+
+    private fun goToLocation(lat: Double, lng: Double) {
+        val latLng = LatLng(lat, lng)
+        val cameraUpdate = CameraUpdateFactory.newLatLngZoom(latLng, 18F)
+        map.moveCamera(cameraUpdate)
+        map.mapType = GoogleMap.MAP_TYPE_NORMAL
     }
 
     private fun onLocationSelected() {
@@ -44,24 +188,148 @@ class SelectLocationFragment : BaseFragment() {
         //  and navigate back to the previous fragment to save the reminder and add the geofence
     }
 
-    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        inflater.inflate(R.menu.map_options, menu)
+    //--------------------------------------------------
+    // Maps Methods
+    //--------------------------------------------------
+
+    private fun initMap() {
+        if (isPermissionGranted()) {
+            val mapFragment = childFragmentManager
+                .findFragmentById(R.id.map) as SupportMapFragment
+            mapFragment.getMapAsync(this)
+            startLocationUpdates()
+        }
     }
 
-    override fun onOptionsItemSelected(item: MenuItem) = when (item.itemId) {
-        // TODO: Change the map type based on the user's selection.
-        R.id.normal_map -> {
-            true
+    /**
+     * Manipulates the map once available.
+     * This callback is triggered when the map is ready to be used.
+     * This is where we can add markers or lines, add listeners or move the camera. In this case,
+     * we just add a marker near Sydney, Australia.
+     * If Google Play services is not installed on the device, the user will be prompted to install
+     * it inside the SupportMapFragment. This method will only be triggered once the user has
+     * installed Google Play services and returned to the app.
+     */
+    override fun onMapReady(googleMap: GoogleMap) {
+        map = googleMap
+
+        // Add a marker in Sydney and move the camera
+		val sydney = LatLng(-34.0, 151.0)
+		map.addMarker(MarkerOptions().position(sydney).title("Marker in Sydney"))
+		map.moveCamera(CameraUpdateFactory.newLatLng(sydney))
+
+//        val latitude = 40.417297
+//        val longitude = -86.892163
+//        val zoomLevel = 18f
+//        val homeLatLng = LatLng(latitude, longitude)
+//        map.addMarker(MarkerOptions().position(homeLatLng))
+//        map.moveCamera(CameraUpdateFactory.newLatLngZoom(homeLatLng, zoomLevel))
+
+//        val overlaySize = 100f
+//        val androidOverlay = GroundOverlayOptions()
+//            .image(BitmapDescriptorFactory.fromResource(R.drawable.android))
+//            .position(homeLatLng, overlaySize)
+//        map.addGroundOverlay(androidOverlay)
+
+        setMapLongClick(map)
+
+        // Put a marker to location that the user selected
+        setPoiClick(map)
+
+        // Add style to the map
+        setMapStyle(map)
+
+        enableMyLocation()
+
+        // TODO: Zoom to the user location after taking his permission
+    }
+
+    private fun setMapLongClick(map: GoogleMap) {
+        map.setOnMapLongClickListener { latLng ->
+            // A Snippet is Additional text that's displayed below the title.
+            val snippet = String.format(
+                Locale.getDefault(),
+                "Lat: %1$.5f, Long: %2$.5f",
+                latLng.latitude,
+                latLng.longitude
+            )
+
+            map.addMarker(
+                MarkerOptions()
+                    .position(latLng)
+                    .title(getString(R.string.dropped_pin))
+                    .snippet(snippet)
+                    .icon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_BLUE))
+            )
         }
-        R.id.hybrid_map -> {
-            true
+    }
+
+    private fun setPoiClick(map: GoogleMap) {
+        map.setOnPoiClickListener { poi ->
+            val poiMarker = map.addMarker(
+                MarkerOptions()
+                    .position(poi.latLng)
+                    .title(poi.name)
+            )
+            poiMarker?.showInfoWindow()
         }
-        R.id.satellite_map -> {
-            true
+    }
+
+    private fun setMapStyle(map: GoogleMap) {
+        try {
+            // Customize the styling of the base map using a JSON object defined
+            // in a raw resource file.
+            val success = map.setMapStyle(
+                MapStyleOptions.loadRawResourceStyle(
+                    requireActivity(),
+                    R.raw.map_style
+                )
+            )
+            if (!success) {
+                Log.e(TAG, "Style parsing failed.")
+            }
+        } catch (e: Resources.NotFoundException) {
+            Log.e(TAG, "Can't find style. Error: ", e)
         }
-        R.id.terrain_map -> {
-            true
+    }
+
+    private fun isPermissionGranted(): Boolean {
+        val fine = ActivityCompat.checkSelfPermission(
+            requireActivity(),
+            Manifest.permission.ACCESS_FINE_LOCATION
+        ) == PackageManager.PERMISSION_GRANTED
+
+        val coarse = ActivityCompat.checkSelfPermission(
+            requireActivity(),
+            Manifest.permission.ACCESS_COARSE_LOCATION
+        ) == PackageManager.PERMISSION_GRANTED
+        return fine && coarse
+    }
+
+    @SuppressLint("MissingPermission")
+    private fun enableMyLocation() {
+        if (isPermissionGranted()) {
+            map.isMyLocationEnabled = true
+        } else {
+            ActivityCompat.requestPermissions(
+                requireActivity(),
+                arrayOf(Manifest.permission.ACCESS_FINE_LOCATION),
+                REQUEST_LOCATION_PERMISSION
+            )
         }
-        else -> super.onOptionsItemSelected(item)
+    }
+
+    override fun onRequestPermissionsResult(
+        requestCode: Int,
+        permissions: Array<String>,
+        grantResults: IntArray
+    ) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
+        // Check if location permissions are granted and if so enable the location data layer.
+        if (requestCode == REQUEST_LOCATION_PERMISSION) {
+            if (grantResults.isNotEmpty() && (grantResults[0] == PackageManager.PERMISSION_GRANTED)) {
+                enableMyLocation()
+            }
+        }
     }
 }
