+++ starter/app/src/main/res/menu/main_menu.xml
@@ -5,8 +5,4 @@
         android:id="@+id/logout"
         android:title="@string/logout"
         app:showAsAction="always" />
-    <item
-        android:id="@+id/show_maps"
-        android:title="@string/maps"
-        app:showAsAction="always" />
 </menu>
